Create README.md
My name is Isaque, I live in the city of São Francisco do Oeste, RN, Brazil. I'm new to programming and I'm in love with the area. I know I have a long way to go but I'll do it with great effort and a smile on my face.